fix: 診断結果一覧画面のデザイン修正
