+++ .env.exapmle
@@ -1,1 +1,5 @@
 NEXT_PUBLIC_GA_ID=
+CLIENT_ID=
+SECRET=
+PRINTER_HOST=
+DEVICE=
