+++ src/app/data/resultSections.ts
@@ -1,0 +1,111 @@
+//診断結果一覧のデータ
+export const resultSections = [
+  {
+    title: '保守的な経済統制派',
+    bgimage:
+      '/backgroundImage/resultDetailBackgroundConservative.png',
+    cards: [
+      {
+        src: '/Types/EACN.jpg',
+        alt: 'EACN',
+        labels: ['EACN', '経済的保守主義者'],
+      },
+      {
+        src: '/Types/PACN.jpg',
+        alt: 'PACN',
+        labels: ['PACN', '進歩的保守主義者'],
+      },
+      {
+        src: '/Types/EACI.jpg',
+        alt: 'EACI',
+        labels: ['EACI', '国際経済的保守主義者'],
+      },
+      {
+        src: '/Types/PACI.jpg',
+        alt: 'PACI',
+        labels: ['PACI', '国際進歩的保守主義者'],
+      },
+    ],
+  },
+  {
+    title: '自由市場主義派',
+    bgimage:
+      '/backgroundImage/resultDetailBackgroundLiberal.png',
+    cards: [
+      {
+        src: '/Types/ELCN.jpg',
+        alt: 'ELCN',
+        labels: ['ELCN', '自由市場的国際主義者'],
+      },
+      {
+        src: '/Types/PLCN.jpg',
+        alt: 'PLCN',
+        labels: ['PLCN', '保守的自由主義者'],
+      },
+      {
+        src: '/Types/ELCI.jpg',
+        alt: 'ELCI',
+        labels: ['ELCI', '自由市場的国際主義者'],
+      },
+      {
+        src: '/Types/PLCI.jpg',
+        alt: 'PLCI',
+        labels: ['PLCI', '保守的国際自由主義者'],
+      },
+    ],
+  },
+  {
+    title: '進歩的国家主義派',
+    bgimage:
+      '/backgroundImage/resultDetailBackgroundProgressive.png',
+    cards: [
+      {
+        src: '/Types/EALN.jpg',
+        alt: 'EALN',
+        labels: ['EALN', '経済的国家自由主義者'],
+      },
+      {
+        src: '/Types/ELLN.jpg',
+        alt: 'ELLN',
+        labels: ['ELLN', '国家的自由主義者'],
+      },
+      {
+        src: '/Types/PALN.jpg',
+        alt: 'PALN',
+        labels: ['PALN', '進歩的国家自由主義者'],
+      },
+      {
+        src: '/Types/PLLN.jpg',
+        alt: 'PLLN',
+        labels: ['PLLN', '進歩的自由国家主義者'],
+      },
+    ],
+  },
+  {
+    title: '国際協力重視派',
+    bgimage:
+      '/backgroundImage/resultDetailBackgroundInternational.png',
+    cards: [
+      {
+        src: '/Types/EALI.jpg',
+        alt: 'EALI',
+        labels: ['EALI', '国際経済的自由主義者'],
+      },
+      {
+        src: '/Types/ELLI.jpg',
+        alt: 'ELLI',
+        labels: ['ELLI', '国際的自由主義者'],
+      },
+      {
+        src: '/Types/PALI.jpg',
+        alt: 'PALI',
+        labels: ['PALI', '国際進歩的自由主義者'],
+      },
+      {
+        src: '/Types/PLLI.jpg',
+        alt: 'PLLI',
+        labels: ['PLLI', '進歩的国際自由主義者'],
+      },
+    ],
+  },
+];
